requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
black==23.9.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
colorama==0.4.6
curlify==2.2.1
deprecation==2.1.0
Flask==2.3.2
frozenlist==1.3.3
functions-framework==3.4.0
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-bigquery==3.12.0
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.4
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.1
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
watchdog==3.0.0
Werkzeug==2.3.5
yarl==1.9.2
